+++ pages/dev_multi_tenant/src/collections/authors.ts
@@ -1,0 +1,40 @@
+import { PageCollectionConfig } from '@jhb.software/payload-pages-plugin'
+
+export const Authors: PageCollectionConfig = {
+  slug: 'authors',
+  admin: {
+    useAsTitle: 'name',
+  },
+  page: {
+    parent: {
+      collection: 'pages',
+      name: 'parent',
+      sharedDocument: true,
+    },
+    breadcrumbs: {
+      labelField: 'name',
+    },
+    slug: {
+      // Disable the slug uniqueness because of the multi-tenant setup (see indexes below)
+      unique: false,
+    },
+  },
+  indexes: [
+    {
+      fields: ['slug', 'tenant'],
+      unique: true,
+    },
+  ],
+  fields: [
+    {
+      name: 'name',
+      type: 'text',
+      required: true,
+    },
+    {
+      name: 'content',
+      type: 'textarea',
+      required: true,
+    },
+  ],
+}