+++ pages/dev_multi_tenant/src/collections/tenants.ts
@@ -1,0 +1,24 @@
+import { CollectionConfig } from 'payload'
+
+const Tenants: CollectionConfig = {
+  slug: 'tenants',
+  admin: {
+    useAsTitle: 'name',
+  },
+  fields: [
+    {
+      name: 'slug',
+      type: 'text',
+      required: true,
+      unique: true,
+      index: true,
+    },
+    {
+      name: 'name',
+      type: 'text',
+      required: true,
+    },
+  ],
+}
+
+export default Tenants